MMVv2: Move InstanceDomainStore closer to InstanceDomain
